Add publication_state column for images
This adds a publication_state for images and modifies the image uploader
to set the initial value for the state. We don't currently have a need
for error states, so for now we have sent_to_draft and sent_to_live.

This was also an opportunity to consolidate the asset manager logic into
the image uploader; as the spec was already covered by the specs for the
classes it delegates to, it seemed better to remove it than to fix it.