Add managing editor permission
This ensures only managing editors can see the 'Retire' and 'Update reason for
retiring' links on a document. We also authenticate for managing editor
permissions in the Retire Documents controller to prevent other users from
accessing retiring by other methods.

There are a couple of places in the view where logic has been added to control
what users without managing editor permissions can see which might seem a
little icky:

- We're only displaying the 'Remove' link for users with managing editor
  permissions at this stage. While this is functionality that hasn't been
implemented yet, we are expecting to do this soon and it didn't seem a good use
of time to temporarily amend the CSS for the 'Remove' link so that it doesn't
align to the right if the 'Retire' link isn't visible.

- Because we are not providing any retirement-related actions for users without
  managing editor permissions, when a document has been retired a bar was still
appearing above the 'View published edition on GOV.UK' link with no additional
links above it, and it looked a bit odd. To get around this logic has been
added to not display this bar in cases where a document has been retired and
the user doesn't have managing editor permission.