Handle Publishing API being down when retiring a document
This ensures that an error message is displayed to users in the event that they
attempt to retire a document but the Publishing API is unavailable and unable
to process the request.

The Document Unpublishing Service wraps retiring a document in a transaction so
if the Publishing API is unavailable then the document's state is unchanged.