Add entry_type of `internal_note` for timeline_entries
This adds the ability to create a timeline_entry with a type of
`internal_note` and display the note on the document history page.

Frontend changes:
I've also swapped the timestamp to come before the internal note /
explanatory note as this groups the entries a little clearer.

I've added a two thirds column to prevent long sentences as per the
prototype design, this also leaves space to add the internal notes
button.

The body of a internal note is convereted from plain text to preserve
line breaks and to create hyperlinks. This is done via `simple_format`
which converts `\n` to `<br>` and a gsub which checks for links and adds
a href accordingly. The whole body of text is wrapped around `sanitize`
which is the safe alternative to `html_safe` as it checks input against a
whitelist and doesn't allow malicous tags / scripts to be rendered.